refactor: service.js -> apiClient.js 파일 이름 변경
